Add typecast to fix compiling in MSVC
BUG=none
TEST=make && make runtests

Reviewed-on: http://gerrit.chromium.org/gerrit/4192
Tested-by: Randall Spangler <rspangler@chromium.org>

+++ firmware/lib/vboot_firmware.c
@@ -300,7 +300,7 @@ int LoadFirmware(LoadFirmwareParams* params) {
       /* Save the good index, now that we're sure we can actually use
        * this firmware.  That's the one we'll boot. */
       good_index = index;
-      shared->firmware_index = index;
+      shared->firmware_index = (uint8_t)index;
       /* Since we now know which firmware to boot, we can update the
        * bootable firmware key block mode. */
       boot_fw_keyblock_flags = key_block->key_block_flags;
